[xds_eds_parsing_move] move ClusterLoadAssignmentBuilder to package client, will unexport later
